Functional Commit 6
This separates out the SpdmTransport trait into a platform module.

This will likely conflict with upstream changes made to support PCI-DOE.
