+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/mod.rs
@@ -11,7 +11,8 @@ use libtock_console::{Console, ConsoleWriter};
 use spdm_lib::codec::MessageBuf;
 use spdm_lib::context::SpdmContext;
 use spdm_lib::protocol::*;
-use spdm_lib::transport::{MctpTransport, SpdmTransport};
+use spdm_lib::transport::MctpTransport;
+use spdm_lib::platform::transport::SpdmTransport;
 use spdm_lib::platform::hash::{SpdmHash, SpdmHashResult, SpdmHashAlgoType, SpdmHashError};
 use spdm_lib::platform::rng::{SpdmRng, SpdmRngError, SpdmRngResult};
 use spdm_lib::platform::evidence::{SpdmEvidence, SpdmEvidenceError, SpdmEvidenceResult};
