+++ runtime/userspace/api/spdm-lib/src/context.rs
@@ -18,7 +18,7 @@ use crate::protocol::version::*;
 use crate::protocol::DeviceCapabilities;
 use crate::state::{ConnectionState, State};
 use crate::transcript::{TranscriptContext, TranscriptManager};
-use crate::transport::SpdmTransport;
+use crate::platform::transport::SpdmTransport;
 use crate::platform::hash::SpdmHash;
 use crate::platform::rng::SpdmRng;
