+++ runtime/userspace/api/spdm-lib/src/platform/mod.rs
@@ -1,3 +1,4 @@
 pub mod hash;
 pub mod rng;
 pub mod evidence;
+pub mod transport;
